Fix warnngs

+++ Core/GanttChartExt/GanttCtrl.cpp
@@ -4182,7 +4182,7 @@ BOOL CGanttCtrl::DrawToday(CDC* pDC, const CRect& rMonth, int nMonth, int nYear,
 	if (bSelected)
 		rToday.DeflateRect(0, 1);
 
-	BYTE cFillOpacity = (rToday.Width() > 1 ? 128 : 255);
+	BYTE cFillOpacity = (BYTE)((rToday.Width() > 1) ? 128 : 255);
 	GraphicsMisc::DrawRect(pDC, rToday, m_crToday, CLR_NONE, 0, GMDR_NONE, cFillOpacity);
 
 	return TRUE;
